Choices flipped for the IFNULL and IFNONNULL and also moved the test in one file
