+++ jpf-symbc/src/examples/npe/StringNullAndNonNullTest.java
@@ -1,0 +1,79 @@
+/*
+ * Copyright (C) 2014, United States Government, as represented by the
+ * Administrator of the National Aeronautics and Space Administration.
+ * All rights reserved.
+ *
+ * Symbolic Pathfinder (jpf-symbc) is licensed under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package npe;
+
+import org.sosy_lab.sv_benchmarks.Verifier;
+
+public class StringNullAndNonNullTest {
+    public static void main(String[] args) {
+        // Tests cases for testing the IFNULL and IFNONNULL bytecodes using symbolic and concrete inputs.
+    }
+
+    // Test 1: IFNULL
+    public static void testIfNull() {
+        String str = Verifier.nondetString();
+        if (str != null) {
+            System.out.println("str is not null");
+        } else {
+            System.out.println("str is null");
+        }
+    }
+
+    // Test 2: IFNONNULL
+    public static void testIfNonNull() {
+        String str = Verifier.nondetString();
+        if (str == null) {
+            System.out.println("str is null");
+        }  else {
+            System.out.println("str is not null");
+        }
+    }
+
+    // Test 3: Multiple checks in sequence
+    public static void testMultipleNullChecks() {
+        String str1 = Verifier.nondetString();
+        String str2 = Verifier.nondetString();
+
+        if (str1 == null) {
+            System.out.println("str1 is null");
+            if (str2 == null) {
+                System.out.println("Both strings are null");
+            } else {
+                System.out.println("str1 null, str2 not null");
+            }
+        } else {
+            System.out.println("str1 is not null");
+            if (str2 == null) {
+                System.out.println("str1 not null, str2 null");
+            } else {
+                System.out.println("Both strings are not null");
+            }
+        }
+    }
+
+    // Test 4: Concrete check
+    public static void testConcrete() {
+        String str = "Hello World";
+        if (str != null) {
+            System.out.println("str is not null");
+        } else {
+            System.out.println("str is null");
+        }
+    }
+}
